+++ internal/colors/colors.go
@@ -3,21 +3,21 @@ package colors
 import pp "github.com/engmtcdrm/go-prettyprint"
 
 // LightGreen returns the input text formatted with a light green color.
-func LightGreen(text string) string {
-	return pp.Fg24Bit(169, 209, 61, text)
+func LightGreen(a ...any) string {
+	return pp.Fg24Bit(169, 209, 61, a...)
 }
 
 // Green returns the input text formatted with a green color.
-func Green(text string) string {
-	return pp.Fg24Bit(97, 179, 79, text)
+func Green(a ...any) string {
+	return pp.Fg24Bit(97, 179, 79, a...)
 }
 
 // MediumGreen returns the input text formatted with a medium green color.
-func MediumGreen(text string) string {
-	return pp.Fg24Bit(62, 139, 69, text)
+func MediumGreen(a ...any) string {
+	return pp.Fg24Bit(62, 139, 69, a...)
 }
 
 // DarkGreen returns the input text formatted with a dark green color.
-func DarkGreen(text string) string {
-	return pp.Fg24Bit(51, 87, 59, text)
+func DarkGreen(a ...any) string {
+	return pp.Fg24Bit(51, 87, 59, a...)
 }